+++ engine/src/main/java/org/terasology/engine/TerasologyEngine.java
@@ -67,7 +67,6 @@ import org.terasology.rendering.nui.skin.UISkin;
 import org.terasology.rendering.nui.skin.UISkinData;
 import org.terasology.version.TerasologyVersion;
 import org.terasology.world.block.family.BlockFamilyRegistry;
-import org.terasology.world.block.family.BlockFamilyRegistryImpl;
 import org.terasology.world.block.loader.BlockFamilyDefinition;
 import org.terasology.world.block.loader.BlockFamilyDefinitionData;
 import org.terasology.world.block.loader.BlockFamilyDefinitionFormat;
@@ -303,7 +302,7 @@ public class TerasologyEngine implements GameEngine {
     }
 
     private void initAssets() {
-        BlockFamilyRegistryImpl familyFactoryRegistry = new BlockFamilyRegistryImpl();
+        BlockFamilyRegistry familyFactoryRegistry = new BlockFamilyRegistry();
         rootContext.put(BlockFamilyRegistry.class, familyFactoryRegistry);
 
         // cast lambdas explicitly to avoid inconsistent compiler behavior wrt. type inference
