+++ engine/src/main/java/org/terasology/world/block/BlockExplorer.java
@@ -16,7 +16,6 @@
 package org.terasology.world.block;
 
 import org.terasology.assets.management.AssetManager;
-import org.terasology.world.block.family.BlockFamilyRegistry;
 import org.terasology.world.block.loader.BlockFamilyDefinition;
 
 import java.util.Optional;
