+++ engine/src/main/java/org/terasology/world/block/family/UpdatesWithNeighboursFamily.java
@@ -21,7 +21,7 @@ import org.terasology.world.block.Block;
 /**
  * Interface for Block family that gets updated by a change in a neighbor block.
  */
-public interface UpdatesWithNeighboursFamily {
+public interface UpdatesWithNeighboursFamily extends BlockFamily{
     /**
      * Update called when a neighbor block changes
      **/
